Add overwrite flag to tar extract
Default behavior for `tar` implementation is to delete the files at the
target location before unpacking the file. This behavior can confuse
certain tools based on inotify, since the inode that they are watching
is deleted as a part of the extract process. Adding `--overwrite`
preserves the inode during the extract allowing watchers to use a simple
inotify watch, without having to specially manage deletions.

The `--overwite` flag is compatible with both GNU and BusyBox tar
implementations.